+++ packages/frontend/src/components/Home/LandingPanel/index.tsx
@@ -2,9 +2,9 @@ import { subject } from '@casl/ability';
 import { Group, Stack, Text, Title } from '@mantine/core';
 import { type FC } from 'react';
 
+import { Can } from '../../../providers/Ability';
 import useApp from '../../../providers/App/useApp';
 import { EventName } from '../../../types/Events';
-import { Can } from '../../common/Authorization';
 import MantineLinkButton from '../../common/MantineLinkButton';
 
 interface Props {