+++ packages/frontend/src/components/PrivateRoute.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, type FC } from 'react';
 import { Navigate, useLocation } from 'react-router';
 import { useEmailStatus } from '../hooks/useEmailVerification';
+import { useAbilityContext } from '../providers/Ability/useAbilityContext';
 import useApp from '../providers/App/useApp';
 import PageSpinner from './PageSpinner';
-import { useAbilityContext } from './common/Authorization/useAbilityContext';
 
 const PrivateRoute: FC<React.PropsWithChildren> = ({ children }) => {
     const {