+++ packages/frontend/src/ee/providers/Embed/EmbedProvider.tsx
@@ -1,8 +1,14 @@
-import { useState, type FC } from 'react';
+import { type FC } from 'react';
 import AppProviderContext from './context';
 
-const EmbedProvider: FC<React.PropsWithChildren<{}>> = ({ children }) => {
-    const [embedToken] = useState(window.location.hash.replace('#', ''));
+type Props = {
+    embedToken?: string;
+};
+
+const EmbedProvider: FC<React.PropsWithChildren<Props>> = ({
+    children,
+    embedToken = window.location.hash.replace('#', ''),
+}) => {
     return (
         <AppProviderContext.Provider value={{ embedToken }}>
             {children}